A complete re-write of the embedded web capability in nodehandler. There is now a web interface in NH with a set of 'tabs' which are coded as plugins. The set of tabs can be found in handler/web/tabs. Each tab has an init.rb file which is loaded by the web server at startup and should install the tab appropriately in the web server.
Also added a few new comman line options. '--web' to allow starting up NH without a experiment file. The idea (not implemented yet) here is that it should be possible to upload an experiment file through the web frontend and execute it. '-D/--debug' short-circuits a few calls which would otherwise call out to gridservices. This was necessary to allow testing outside a testbed.

There are still a few reamining hard-coded shortcuts (this code was developed for the GEC4 demo) which need to be cleaned up. For instance, the web server needs to figure out what address it is being called on to set the right URL for the flash-based visualization widget which needs to call back for configuration parameters. In addition, the 'public_html' directory should really move out of the repository. Maybe the supporting files should by default be accessible from the mytestbed.net site. It's also unclear to me right now from where the flash file is served from. It doesn't seem to show up in the access logs.