Major reworking of the visualisation support. Broke the Arel support into a frontend and a backend. The frontend, which can be used by any application is essentially just creating an XML document which is posted to the new 'result2' AM service where the actual database query is performed. Results are shipped back via XML in the HTTP reply.
Please note that right now, little error checking is performed. For instance if the POST to the AM fails, no proper
error message is returned to the web frontend.

The Arel library has also been removed from the package as it is no longer used - anyway the gem is now much further 
along.